+++ packages/desktop/src/footer-nav/FooterNav.tsx
@@ -5,17 +5,26 @@ import {Box, Flex, FlexItem} from '@qiwi/pijma-core'
 import {NavLink, NavLinkProps} from '../nav-link'
 
 export interface FooterNavProps {
-  children: NavLinkProps[]
+  children?: NavLinkProps[]
+  stub?: Array<number> | boolean
 }
 
-export const FooterNav: FC<FooterNavProps> = ({children}) => (
+export const FooterNav: FC<FooterNavProps> = ({children = [], stub = false}) => (
   <Box overflow="hidden">
     <Flex wrap="wrap" mx={-2} my={-2.5}>
-      {children.map((item, i) => (
-        <FlexItem px={2} py={2.5} key={i}>
-          <NavLink {...item}/>
-        </FlexItem>
-      ))}
+      {stub ? (
+        (typeof stub === 'boolean' ? [32, 40, 26] : stub).map((width, index) => (
+          <FlexItem px={2} py={2.5} key={index} width={width}>
+            <NavLink stub/>
+          </FlexItem>
+        ))
+      ) : (
+        children.map((item, i) => (
+          <FlexItem px={2} py={2.5} key={i}>
+            <NavLink {...item}/>
+          </FlexItem>
+        ))
+      )}
     </Flex>
   </Box>
 )