+++ packages/desktop/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@qiwi/pijma-desktop-v1.102.0](https://github.com/qiwi/pijma/compare/v1.101.1...v1.102.0) (2020-01-31)
+
+
+### Features
+
+* stub footer-nav ([e3eb002](https://github.com/qiwi/pijma/commit/e3eb002))
+
+# [@qiwi/pijma-desktop-v1.101.1](https://github.com/qiwi/pijma/compare/v1.101.0...v1.101.1) (2020-01-31)
+
+
+### Bug Fixes
+
+* stub text-field ([caaddbb](https://github.com/qiwi/pijma/commit/caaddbb))
+
+# [@qiwi/pijma-desktop-v1.101.0](https://github.com/qiwi/pijma/compare/v1.100.0...v1.101.0) (2020-01-31)
+
+
+### Features
+
+* **desktop:** stub header-menu ([aa7fb6d](https://github.com/qiwi/pijma/commit/aa7fb6d))
+
 # [@qiwi/pijma-desktop-v1.100.0](https://github.com/qiwi/pijma/compare/v1.99.0...v1.100.0) (2020-01-29)